OPEN-104: Refactored apdu tools to a new module
