+++ core-modules/healthcard-apdu-base/Cargo.toml
@@ -19,18 +19,16 @@
 # For additional notes and disclaimer from gematik and in case of changes by gematik,
 # find details in the "Readme" file.
 
-[workspace]
-members = [
-    "core-modules/asn1",
-    "core-modules/crypto",
-    "core-modules/crypto-openssl-sys",
-    "core-modules/healthcard",
-    "core-modules/healthcard-apdu-base",
-    "core-modules/healthcard-apdu-tools",
-    "tools/uniffi-cli",
-]
-resolver = "2"
-[workspace.package]
+[package]
+name = "healthcard-apdu-base"
 version = "0.1.0"
-edition = "2024"
+edition = "2021"
 license = "Apache-2.0"
+
+[dependencies]
+crypto = { path = "../crypto" }
+thiserror = "2.0.16"
+hex = "0.4.3"
+num-bigint = "0.4"
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.117"
