+++ core-modules/healthcard-apdu-tools/Cargo.toml
@@ -19,18 +19,20 @@
 # For additional notes and disclaimer from gematik and in case of changes by gematik,
 # find details in the "Readme" file.
 
-[workspace]
-members = [
-    "core-modules/asn1",
-    "core-modules/crypto",
-    "core-modules/crypto-openssl-sys",
-    "core-modules/healthcard",
-    "core-modules/healthcard-apdu-base",
-    "core-modules/healthcard-apdu-tools",
-    "tools/uniffi-cli",
-]
-resolver = "2"
-[workspace.package]
+[package]
+name = "healthcard-apdu-tools"
 version = "0.1.0"
-edition = "2024"
+edition = "2021"
 license = "Apache-2.0"
+
+[features]
+default = []
+pcsc = ["dep:pcsc"]
+
+[dependencies]
+healthcard-apdu-base = { path = "../healthcard-apdu-base" }
+healthcard = { path = "../healthcard", default-features = false }
+crypto = { path = "../crypto" }
+hex = "0.4.3"
+clap = { version = "4.5.40", features = ["derive"] }
+pcsc = { version = "2.8.2", optional = true }
