+++ core-modules/healthcard/Cargo.toml
@@ -31,9 +31,7 @@ name = "healthcard"
 
 [features]
 default = ["uniffi"]
-ci = ["uniffi", "apdu-tools"]
-apdu-tools = ["dep:clap", "dep:serde", "dep:serde_json"]
-pcsc = ["dep:pcsc"]
+ci = ["uniffi"]
 
 [dependencies]
 lazy_static = "1.5.0"
@@ -45,11 +43,6 @@ once_cell = "1.21.3"
 num-bigint = "0.4"
 zeroize = { version = "1.8.1", features = ["zeroize_derive"] }
 uniffi = { version = "0.30.0", optional = true }
-clap = { version = "4.5.40", optional = true, features = ["derive"] }
-serde = { version = "1.0.219", optional = true, features = ["derive"] }
-serde_json = { version = "1.0.117", optional = true }
-pcsc = { version = "2.8.2", optional = true }
 
 [dev-dependencies]
-serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.117"
+healthcard-apdu-base = { path = "../healthcard-apdu-base" }
